--- publication-commons/src/main/java/no/unit/nva/JacocoGenerated.java
@@ -1,5 +1,0 @@
-package no.unit.nva;
-
-// Named JacocoGenerated because Jacoco ignores methods annotated with name which contains Generated
-public @interface JacocoGenerated {
-}
