--- publication-commons/src/main/java/no/unit/nva/Logger.java
@@ -1,16 +1,0 @@
-package no.unit.nva;
-
-import java.util.Arrays;
-
-@JacocoGenerated
-public class Logger {
-    
-    public static void logError(Throwable e) {
-        System.out.println(Arrays.asList(e.getStackTrace()));
-    }
-
-    public static void log(String message) {
-        System.out.println(message);
-    }
-
-}
