+++ publication-commons/src/main/java/no/unit/publication/JacocoGenerated.java
@@ -1,0 +1,5 @@
+package no.unit.publication;
+
+// Named JacocoGenerated because Jacoco ignores methods annotated with name which contains Generated
+public @interface JacocoGenerated {
+}
