+++ publication-commons/src/main/java/no/unit/publication/Logger.java
@@ -1,0 +1,16 @@
+package no.unit.publication;
+
+import java.util.Arrays;
+
+@JacocoGenerated
+public class Logger {
+    
+    public static void logError(Throwable e) {
+        System.out.println(Arrays.asList(e.getStackTrace()));
+    }
+
+    public static void log(String message) {
+        System.out.println(message);
+    }
+
+}
