+++ publications-by-owner/src/main/java/no/unit/nva/publication/owner/PublicationsByOwnerHandler.java
@@ -4,14 +4,14 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import no.unit.nva.Environment;
-import no.unit.nva.GatewayResponse;
-import no.unit.nva.JacocoGenerated;
-import no.unit.nva.PublicationHandler;
-import no.unit.nva.model.PublicationSummary;
+import no.unit.publication.Environment;
+import no.unit.publication.GatewayResponse;
+import no.unit.publication.JacocoGenerated;
+import no.unit.publication.PublicationHandler;
+import no.unit.publication.model.PublicationSummary;
 import no.unit.nva.model.util.OrgNumberMapper;
-import no.unit.nva.service.PublicationService;
-import no.unit.nva.service.impl.DynamoDBPublicationService;
+import no.unit.publication.service.PublicationService;
+import no.unit.publication.service.impl.DynamoDBPublicationService;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -20,8 +20,8 @@ import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 
-import static no.unit.nva.Logger.log;
-import static no.unit.nva.Logger.logError;
+import static no.unit.publication.Logger.log;
+import static no.unit.publication.Logger.logError;
 import static org.zalando.problem.Status.BAD_GATEWAY;
 import static org.zalando.problem.Status.BAD_REQUEST;
 import static org.zalando.problem.Status.INTERNAL_SERVER_ERROR;
