+++ publications-by-owner/src/test/java/no/unit/nva/publication/owner/PublicationsByOwnerHandlerTest.java
@@ -2,11 +2,11 @@ package no.unit.nva.publication.owner;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import no.unit.nva.Environment;
-import no.unit.nva.GatewayResponse;
-import no.unit.nva.PublicationHandler;
-import no.unit.nva.model.PublicationSummary;
-import no.unit.nva.service.PublicationService;
+import no.unit.publication.Environment;
+import no.unit.publication.GatewayResponse;
+import no.unit.publication.PublicationHandler;
+import no.unit.publication.model.PublicationSummary;
+import no.unit.publication.service.PublicationService;
 import org.apache.http.HttpHeaders;
 import org.apache.http.entity.ContentType;
 import org.junit.Assert;
@@ -29,8 +29,8 @@ import java.util.Optional;
 import java.util.UUID;
 
 import static java.util.Collections.singletonMap;
-import static no.unit.nva.PublicationHandler.ACCESS_CONTROL_ALLOW_ORIGIN;
-import static no.unit.nva.PublicationHandler.ALLOWED_ORIGIN_ENV;
+import static no.unit.publication.PublicationHandler.ACCESS_CONTROL_ALLOW_ORIGIN;
+import static no.unit.publication.PublicationHandler.ALLOWED_ORIGIN_ENV;
 import static no.unit.nva.model.PublicationStatus.DRAFT;
 import static org.apache.http.HttpHeaders.CONTENT_TYPE;
 import static org.apache.http.HttpStatus.SC_BAD_GATEWAY;
